// api/models/SchedulerStatistics.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 返回对象
 */
export type SchedulerStatistics = {
    pendingTasks?: number;
    runningTasks?: number;
    completedTasks?: number;
    failedTasks?: number;
    onlineVolunteers?: Record<string, any>;
};